chore: rename mocks folder
